feat: rename utils to helpers and move media-query there
