+++ src/helpers/index.ts
@@ -1,2 +1,3 @@
 export type { PolymorphicRef, PolymorphicComponentProps } from './types';
+export { default as mediaQuery } from './media-query';
 export { default as hexToRgba } from './hex-to-rgba';
