--- src/media-query/index.ts
@@ -1,1 +1,0 @@
-export { default as mediaQuery } from './media-query';
